backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
bcrypt
PyJWT
pydantic
scikit-learn
numpy
beautifulsoup4
requests
PyPDF2
python-docx
python-multipart
